Configure for three-stage SDP; from ROM, then from SPL, then from U-Boot.
The ROM starts the Serial Download Protocol when configured or when it
cannot boot from it's boot medium. Once SPL has started, it recognizes the
SDP mode and will re-enumerate again and wait for SDP. In U-Boot, the
'sdp' command will re-enumerate once more and wait for SDP. Note that
besides transferring files into RAM, also scripts can be transferred and
run.

The whole approach is meant for manufacturing purposes (and debricking)
where the boot medium is empty, and must be boot strapped over USB.

Signed-off-by: Leon Woestenberg <leon@sidebranch.com>